+++ packages/plugin-tailwind/vitest.config.ts
@@ -2,6 +2,6 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    testTimeout: 10_000,
+    testTimeout: 15_000,
   },
 });